Porting Really Slick Screensavers
This pulls in a port of the GLX version of Really Slick Screensavers.

This has three of the live wallpapers, and allows settings for each.

They were not originally meant for Android, so some changes were
necessary.  Particularly with regards to Android Service lifecycles as
opposed to standard desktop OS screensavers which are either running or
not on one screen.  So I worked towards making it more robust - reducing
state, encapsulating, putting mutual exclusion around critical sections
(mainly draw calls), replacing getopt with the more thread-safe optlist
etc.  That work is not yet done, but it is less crash-prone now than in
its original ported state.

The RSS-GLX code uses OpenGL, not Android's OpenGL ES.  I used the shim
that was in Xscreensaver to do the translation of OpenGL to OpenGL ES
calls.  Including a partially implemented glLoadMatrixf call which I
wrote (but did not finish).